remove circular dependency in state
